Implement hooks to control filesystem mounting
The original version of main() would handle mounting and unmounting; but that meant that
there wasn't a clean way to remount the filesystems, which is the whole reason for this project!

So, instead I added a hook object that providers of main() will need to provide.

If a main() provider defines a hook object with no entries, it will use the default behavior of
mounting the SD card (and USB if ioshax is available).

A main() provider must provide both the mount() callback and unmount() callback.

Also, added a placeholder package XML + PNG.

== TESTING
I actually put this on my wiiu and.. well, it didn't crash, but it also exits immediately,
which is probably my own dumb mistake. Will fix later.