Add network logging to UDP, a little more console output
So, the buffer code isn't quite functional--it successfully detects the newline, but
the command is an empty string.

To aid in debugging this, I've brought in the UDP logging I put into RetroArch so
I can get complete debug info (the on-screen console chops off at 80 chars and
has no backscroll).

On the plus side: I verified the debounce works!